requirements.txt
langchain
langchain_community
langchain_core
langgraph
python-dotenv
langchain-google-genai
google-generativeai
sqlalchemy
typing_extensions
flask